Added placeholder documentation.html
For now, this file should be the documentation for using the program as
well as our own documentation for how it works. Additionally, we could
keep some interesting history and things here.

Maybe we could eventually split that into a few separate documents, but
for now I figured we could just start with this and get rolling!